+++ springcore-20190324/springcore-20190324.iml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module version="4">
+  <component name="FacetManager">
+    <facet type="web" name="Web">
+      <configuration>
+        <descriptors>
+          <deploymentDescriptor name="web.xml" url="file://$MODULE_DIR$/../springcore-20190324/src/main/webapp/WEB-INF/web.xml" />
+        </descriptors>
+        <webroots>
+          <root url="file://$MODULE_DIR$/../springcore-20190324/src/main/resources" relative="WEB-INF/classes" />
+          <root url="file://$MODULE_DIR$/../springcore-20190324/src/main/webapp" relative="/" />
+        </webroots>
+        <sourceRoots>
+          <root url="file://$MODULE_DIR$/../springcore-20190324/src/main/java" />
+          <root url="file://$MODULE_DIR$/../springcore-20190324/src/main/resources" />
+        </sourceRoots>
+      </configuration>
+    </facet>
+  </component>
+</module>
