+++ springcore-20190324/src/main/java/springcore01/classcode/com/gupaoedu/demo/mvc/action/TestAction.java
@@ -1,0 +1,5 @@
+package springcore01.classcode.com.gupaoedu.demo.mvc.action;
+
+
+public class TestAction {
+}
