+++ springcore-20190324/src/main/java/springcore01/classcode/com/gupaoedu/demo/service/IDemoService.java
@@ -1,0 +1,8 @@
+package springcore01.classcode.com.gupaoedu.demo.service;
+
+
+public interface IDemoService {
+	
+	String get(String name);
+	
+}
