+++ springcore-20190324/src/main/java/springcore01/classcode/com/gupaoedu/mvcframework/annotation/GPController.java
@@ -1,0 +1,10 @@
+package springcore01.classcode.com.gupaoedu.mvcframework.annotation;
+
+import java.lang.annotation.*;
+
+@Target({ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+public @interface GPController {
+    String value() default "";
+}
