+++ springcore-20190324/src/main/resources/system.properties
@@ -1,0 +1,3 @@
+#************** System Config *******************#
+
+system.charSet=UTF-8
