--- springcore-20191215/src/main/java/springcore01/classcode/com/gupaoedu/demo/mvc/action/TestAction.java
@@ -1,5 +1,0 @@
-package springcore01.classcode.com.gupaoedu.demo.mvc.action;
-
-
-public class TestAction {
-}
