--- springcore-20191215/src/main/java/springcore01/classcode/com/gupaoedu/demo/mvc/action/TwoAction.java
@@ -1,23 +1,0 @@
-package springcore01.classcode.com.gupaoedu.demo.mvc.action;
-
-import springcore01.classcode.com.gupaoedu.demo.service.IDemoService;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class TwoAction {
-	
-	private IDemoService demoService;
-
-	public void edit(HttpServletRequest req,HttpServletResponse resp,
-					 String name){
-		String result = demoService.get(name);
-		try {
-			resp.getWriter().write(result);
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
-	}
-	
-}
