--- springcore-20191215/src/main/java/springcore01/classcode/com/gupaoedu/demo/service/IDemoService.java
@@ -1,8 +1,0 @@
-package springcore01.classcode.com.gupaoedu.demo.service;
-
-
-public interface IDemoService {
-	
-	String get(String name);
-	
-}
