--- springcore-20191215/src/main/java/springcore01/classcode/com/gupaoedu/mvcframework/annotation/GPAutowired.java
@@ -1,10 +1,0 @@
-package springcore01.classcode.com.gupaoedu.mvcframework.annotation;
-
-import java.lang.annotation.*;
-
-@Target({ElementType.FIELD})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-public @interface GPAutowired {
-    String value() default "";
-}
