--- springcore-20191215/src/main/java/springcore01/classcode/com/gupaoedu/mvcframework/annotation/GPRequestMapping.java
@@ -1,10 +1,0 @@
-package springcore01.classcode.com.gupaoedu.mvcframework.annotation;
-
-import java.lang.annotation.*;
-
-@Target({ElementType.TYPE,ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-public @interface GPRequestMapping {
-    String value() default "";
-}
