--- springcore-20191215/src/main/java/springcore01/classcode/com/gupaoedu/mvcframework/annotation/GPService.java
@@ -1,10 +1,0 @@
-package springcore01.classcode.com.gupaoedu.mvcframework.annotation;
-
-import java.lang.annotation.*;
-
-@Target({ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-public @interface GPService {
-    String value() default "";
-}
