--- springcore-20191215/src/main/resources/system.properties
@@ -1,3 +1,0 @@
-#************** System Config *******************#
-
-system.charSet=UTF-8
